Fix schema generation and tensor value-type handling
Schema builder (verified against the Vespa schema reference):
- DistanceMetric.Geodesic now emits "geodegrees" — Vespa has no metric
  named "geodesic", so deployment always failed.
- RankType.Identity/Tags now emit the rank-type element; field-level
  rank: accepts only filter/normal, so rank: identity/tags failed
  deployment.
- When a type has tensor fields and a user-declared [VespaRankProfile
  ("default")], the tensor inputs are merged into the user's profile
  instead of emitting a second rank-profile default block (duplicate
  profile names fail deployment).

Tensors:
- int8 cell values are read as doubles and cast: Vespa renders cell
  values with decimals (1.0), which Utf8JsonReader.GetSByte() rejects.
- bfloat16 now maps to float instead of System.Half: bfloat16 has
  float32 range while Half saturates at 65504, so large-magnitude
  values silently became +∞.

+++ src/Vespa.NET.Models/Tensors/VespaTensor.cs
@@ -5,7 +5,7 @@ namespace Vespa.Models.Tensors;
 
 /// <summary>
 /// Universal representation of a Vespa tensor that supports all formats and value types
-/// Supports: float, double, int8 (sbyte), bfloat16 (Half)
+/// Supports: float, double, int8 (sbyte), bfloat16 (stored as float — lossless)
 /// </summary>
 public sealed class VespaTensor
 {
@@ -119,8 +119,10 @@ public sealed class VespaTensor
             return typeof(double);
         if (typeSpec.Contains("<int8>", StringComparison.Ordinal))
             return typeof(sbyte);
+        // bfloat16 has float32 range (±3.4e38); System.Half (binary16) saturates at
+        // 65504, so float is the lossless in-memory representation.
         if (typeSpec.Contains("<bfloat16>", StringComparison.Ordinal))
-            return typeof(Half);
+            return typeof(float);
 
         return typeof(double); // Default fallback
     }
